backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
pandas>=1.5.0
fairlearn>=0.9.0
scikit-learn>=1.3.0
google-generativeai>=0.3.0
python-dotenv==1.0.1
pydantic>=2.0.0
reportlab>=4.0.0
Pillow>=10.0.0
numpy>=1.23.0
matplotlib>=3.7.0
firebase-admin>=6.5.0
